day 1 gets
This was my first time in a long time touching clojure so some hiccoughs
to be expected.

The main issues I had today were; that my repl was busted so I ended up
copy/pasting between panes a lot to run my changes and the differences
in flow control. I understood the problem, mostly, but got completey
lost in the sauce with first/second/nth and typecasting got me over a
barrell.

4h completion time, fingers crossed day 2 goes quicker